noise_reduction: fix and integrate

+++ noise_reduction/noise_reduction.h
@@ -1,11 +1,11 @@
 #ifndef NOISE_REDUCTION_H
 #define NOISE_REDUCTION_H
 
-#include "../bitmap/bmp.h"
+#include <stdlib.h>
+
 #include "../bitmap/img.h"
 
-
-int ibw(int a, int b, int c);
+size_t ibw(size_t a, size_t b, size_t c);
 t_bw_img* reduce_noise(t_bw_img *img);
 
 #endif
